+++ packages/desktop/renderer/@types/preload.d.ts
@@ -1,9 +1,0 @@
-import { Preload } from '../main/preload'
-
- 
-declare global {
-    interface Window {
-        Blacklight: typeof Preload;
-    }
-}
-