+++ packages/desktop/tests/lib/auth.ts
@@ -1,9 +1,0 @@
-export default {
-    isAuthenticated: async () => {
-
-    },
-
-    performAuthentication: async () => {
-        
-    },
-}